feat: rewrite of lib-sepa
This library has been rewritten.
The expectation is that this requires some tweaks to ensure it behaves correctly, but this should be a good first version to go with.
Next thing will be to create a GitHub action to run a CI pipeline that can publish the library.
Then we need some unit tests to verify proper behaviour of the implemented business logic.